requirements.txt
Pillow
pip install pillow